// src/modules/user/constants/user.interface.ts
import { IWebsite } from 'src/modules/website/constants/website.interface';

export interface IRole {
  id: number;
  name: string;
  slug: string;
}

export interface IUser {
  id: number;
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  website: IWebsite;
  roles: IRole[];
}